Woohoo! Should have looked at prxtool sooner - now the ElfFile loader properly relocates data segments, and as such real games (kind of) run! I can get pretty far in PB, and the trace looks legit!
Fixed a bunch of small things in the IoFileMgrForUser module, as well as some path stuff in the media components. With these changes, including some UmdUser stuff, the pspsdk fileio sample runs and can dump the BOOT.BIN from the UMD (the others don't run because I don't emulate the flash and I don't do the memory ranges in a compatible format for the dump, but that should work too).

Re-enabled the GamePicker (in place of the XMB), as I think it's the future. It's not polished yet, and the primary mode ('Recent Games') isn't implemented, but it does work.