A lot of work on the compatibility stuff - database and sprocs are pretty much done (at least the stuff concerning games), and the web presentation and basic administration (approving and denying additions) is in. Player has a submit game dialog that works, but no login/etc so it isn't ready. A lot of work remains, but it's looking nice.
I'm going to shift in to working on the emulator now, as the class I was doing the database stuff for is over and I really want to get something running. I'm thinking about giving up on XNA and using OpenGL (via Tao), but I need to get a little further before making a final decision.