Little app to help build instruction mix lists (pipe in psp-objdump output or read it from a file and it will spit out a sorted list of instructions and their usage count). It would also be nice to have some statistics stuff in the dynarec cpu to help track down bugs.
Fixed a bug in mult instruction.

Direct3DM video component now runs on MDX1.1, although the framebuffer is messed up and only the triangles in Lesson1 draw.

Still need to find a nice way to test things, as there is obviously a bug in some instruction and it's proving incredibly difficult to find.