Part 2 of refactor. Greatly simplified the component model - there can be only one input plugin, one memory stick plugin, and one umd plugin. They have all been split in to their own members in params and instance. This makes things easier to code against (no iterating over IO to find something, etc). One can argue that having multiple of each would be nice, however I think it would be the exception rather than the rule, and the emulator codebase is already confusing enough. Note that IO and IIODriver still remain for the day when networking/USB/GPS/whatever is added.
Next step is to clean up the CPU interfaces. I'd like to try to move anything that could be in the debug interfaces to the debug interfaces. This is in preparation for the ultra CPU core that I'm getting the itch to work on (the .NET based dynarec will never be fast enough to 100% games - at least not for a few years). I'd also like to see about moving most of the UI code in to interfaces and moving the UI to a separate assembly. I think that enough work has been invested in Player as a host that I don't want to duplicate it if I decide to write a different frontend (and I am thinking about it - the current one is buggy and not so intuitive).

After that I will either work on the ultra CPU or really go in heavy on one of the video plugins (still debating between OGL and MDX). I want to play games, damnit!