Some minor API changes (to help ultra CPU, mainly). One of the biggies is InternalMemoryPointer, which gives an IntPtr when InternalMemory is null but InternalMemorySupported is true (confusing I know...). This issue is that InternalMemory returns a byte[], but since the memory in the ultra CPU is a malloc'ed block of byte*, there is no way to return a byte[]. One other change is in regards to the semantics of ICpuCore GeneralRegisters - it now can be assumed to return a copy, whether or not it really does. Use SetGeneralRegister to set a register.
Fixed some memory bugs in dynarec CPU - they were pretty bad, but didn't seem to affect anything.

Ultra CPU implemented, except for the implementation ;) Everything is in except for actual execution (includes coprocessors and everything). The coprocessor model is the same as in the dynarec CPU, and that means it's bad (uses managed arrays to store registers and such, has nasty control code, etc). It needs to be rewritten.

The next step is to start building the architecture for the ultra CPU (code cache, generator, etc). I think I'm going to build it all with debugging support, but not actually implement it yet.