Fix to make SoftWire SSE instructions public.
FPU (COP1) mainly implemented. Everything but the main instruction handling logic and branch instructions is in.
I wrote it all for SSE, so now Ultra is testable and tests to ensure that the CPU supports SSE2. There is a #define SSE at the top of R4000Generator_Fpu.cpp, however very few of the alternate paths (using x87) are implemented. Almost all methods use SSE, however there were a few that were easier with the FPU. Since I don't know the performance of SSE (using single scalar ops exclusively) vs. x87, I did implement a few that I think may be faster the old way. Later on I can profile FPU heavy code and see what performs best. It may turn out that x87 is always faster because of my usage pattern - if I was using register allocation and fancy stuff SSE would probably be faster (as there are 8 or so SSE registers as opposed to a stack that would be a pain to manage), so the speedup may come some day.