The rest of the instructions (MFCz/MTCz, etc), as well as the code in R4000AdvancedBlockBuilder to support them. Instead of having switches I did lookup tables. Unfortunately cause I'm lazy the tables are 64 elements large when they only need to be about 8. I obviously don't mind a little over-allocation, so who cares ;)
Thought about FCOMPARE - it may be possible to speed it up. It currently does this:
- Load SSE registers with 2 values
- Run the operator
- <SSE register now contains FFFFFFFF is operator true, else 0>
- Move SSE register to GPR
- Trim GPR so 0 or 1
- Store condition bit
My thought is that this is bad because it relies on explicit communication of the register value from the XMM registers to the GPR registers, possibly creating some unnecessary dependencies. The other way I was thinking:
- Load SSE registers with 2 values
- Run COMISS (compare and set EFLAGS)
- SETcc based on desired operator
- Store condition bit
A lot less instructions, and no need to retreive a value from the XMM registers.

If FCOMPARE is frequently called, I'll definitely try making this change.

Speaking of all this perf stuff, I will need to start thinking of how I want to profile things. I'd like to be able to generate runtime instruction execution counts, as well as allow for some in-depth profiling of instructions (like how long each instruction takes to execute, on average, etc). I think I can do something like this by just emitting the code from the block builder, as all instruction groups are atomic. If I was doing any sort of register allocation, this wouldn't work. If anything, it means that if I do decide to do register allocation, I'll have to have a nice way to disable it so that I can gather statistics in the future.