By tweaking MAXCODELENGTH, basic controller sample and empty test run (basic controller tries to set the COP1 control register, which asserts, but it isn't needed), which means others probably run too.
Generation of large blocks is SLOW. There are several blocks that are over 1000 instructions. Since I lost the debug info for these apps, I'm not sure what the hell is happening. I'm going to fix things so blocks will have a maximum size (or at least set a max that it aborts, and hope the code never runs that far!).

The apps themselves are also not running as fast as I had hoped. I think the primary reason is that my jump thunks aren't getting used at all - just a bunch of jump rets (49k vs 17 in quick run I did). If I can keep things in native code as much as possible, I think it'll run a lot better.