Fixes to support inline jump blocks at the end of blocks.
Manually tweaking MAXCODELENGTH now - 200 seems to work for most cases. There is a chunk of code, from gcc or something, that my block builder cannot handle - this causes the jumps inside of it to cause it to generate several times - it hits the max each time, slowing startup down. Will need to investigate and see if I can fix it (or at least hack around it).

I got Lesson1 running, which is the simple triangle drawer. It doesn't draw the square, for some reason, but it does give me enough info to do some perf analysis. For some reason, though, the number of code blocks executed is orders of magnitude higher than the number of jumps - not sure why. What I've seen, though, is that there is a 3x speedup in Lesson1, while in debug, with release running about about 80MIPS. Things would actually run faster if I disabled statistics, which take up a lot of code right now.

At this point I think I'm going to build some better stat gathering about syscalls so I can identify potential candidates for native support. After that, I need to build some more samples to get some things working, and then finally rewrite the whole video driver interface. What's there will work for managed code, but for unmanaged code I need a way for the CPU to communicate with the video driver directly, in an unmanaged fashion. This will of course mean moving the code from the BIOS to the CPU, as well as using some native shared memory buffer for the command list. I'm not concerned with keeping things so nice now :)

I'm also debating whether or not to deprecate the dynarec CPU. With all the new architecture changes, I think the Ultra CPU may be the way to go. I'm sure I can backport a lot of my new techniques in to the dynarec (for jump thunks and inline tail jumps, etc) and get some speedups, however for the next few years I doubt it will be as fast as the ultra is right now.