Added the ability to override BIOS methods. Because this is a tricky process, you must add a BiosStubOverridableAttribute to the methods the BIOS author wants to allow overriding on. This is because some methods may depend on or modify module/BIOS state and must be done there. For methods that are fairly atomic with respect to the BIOS (like RTC and GE stuff), they can be overridden.
The first two I've done are the rtc get frequency and get tick count - these use the Windows API, as using rdtsc is a bad idea on multi-core/proc machines. So now for these methods, no native/managed transition is required, and it's pretty fast (as the code simply jumps to the function and returns the value back).

There is also a new statistic to observe this: BiosSyscallCount will show the # of calls made to the BIOS (requiring the managed/native switch), and NativeSyscallCount will show the # made to native methods implemented in the CPU. Since the number of methods implemented in so small (and will hopefully remain that way), they all live in R4000BiosStubs.cpp for now.

One thing I'm thinking, now that I've finished all this, is that I should have called it BiosStubAtomicAttribute, and defined it as saying no state is changed anywhere. This would mean that after a syscall is executed, unlike now which causes a block to end and control to switch to the main loop, we could stay within native code. Probably going to change all of this now :)