Added some options: - SAFEARITHMETIC in R4000Generator_Arithmetic.cpp makes sure some arithmetic ops are right (mainly shifts) - IGNOREFRAMEBUFFER ignores writes to the framebuffer - they are expensive because they require a transition, and since I don't have a framebuffer implemented in the OGL driver I might as well skip them!
Fixed a major bug in INS/EXT - they had 'long', which came from when I copy/pasted the C# dynarec code - well, of course, in C# long is 64 bits but in C++ it's only 32. Bug! This was causing one of the few functions that uses it, related to 64-bit -> double conversion, to fail, causing anything relating to timing to fail.

With this, Sample 5 (NeHe lesson 3) runs, for real, perfectly. Timed properly and everything!

Added dumpsrc.sh - this is like dumpelf.sh, only it will also inline source if it can. Add -g to CFLAGS to have app source included.

Now emit INT 3 on BREAK - some __udivwhatever code was emitting it.

One thing I'd like to do is be able to detect code patterns and replace them. For example, there are several hundred expensive MIPS instructions (meaning several thousand x86 ones) executed just to convert a 64-bit integer to a double - this operation is 2 instructions in native x86. I'd have to use some sort of static analysis to figure out where the blocks are, and there may be issues with different compilers/optimization levels outputting different versions of it. Would be cool, though.