MAJOR BREAKING CHANGES
I found out some things... things I should have checked a LONG time ago. Turns out that the calling convention used is NOT the standard I was used to (or read about). I had done everything to support the first 4 arguments going in a0-a3, and the rest spilling in to the stack. Well, in reality its the first 4 in a0-a3, the next 4 in t0-t3, then the rest on the stack. Whoops. There is also the issue of the BIOS methods not supporting 64 bit returns, or 64 bit parameters. Double whoops.

So basically, the whole BIOS is wrong and needs to be thrown away. At least the interface. So once I realized this I took it in stride: an excuse to do things right!

A new BIOS, FastHLE, has been added. It is in C++/CLI (now that I'm used to it I love it), and will support the BiosApi native interface. Since it's still managed I can use the great .NET BCL, while still allowing for native calls. This will be critical for things like the synchronization primitives/etc. Without doing this, I fear I would have ended up putting most of the BIOS in the CPU! In fact, now that the native interface exists, I can undo some of the stuff I did in R4000BiosStubs.cpp and bring the syscalls back in to the BIOS. Some will stay (particularly the inlined ones), but I may even move the video interface out - it really belongs in the BIOS, and there won't be that much of a hit by moving it there (since the new BIOS can use the native video interface too).

Because of the radical changes required to support the newfound calling convention, there is no way to really salvage the existing BIOS. It will be dropped as soon as I move its good code in to the new BIOS (although I do plan to rewrite a lot to make it cleaner).

In stark contrast to the old BIOS call interface (where every BIOS method had the same delegate and a lot of ugly work was required) the new interface is much more natural. Two new helpers (of which only the managed version is implemented) will emit the proper code to call the target method by setting up the arguments and handling the return. Right now the method, EmitShim, lives in R4000Cpu.cpp. It actually emits MSIL and takes no parameters and returns nothing, and is designed to do everything required to invoke the BIOS call it wraps. It's pretty neat trick, and will hopefully help performance with managed calls. The added advantage is that it builds itself to support the target method, as opposed to before where the target method had to be formatted to match the calling convention. As soon as I get things working I'll worry about the native BIOS interface - it won't need a special shim method so much as just having the right code emitted on a SYSCALL instruction.

GenerateStubs was retarded. It's gone now, and in its place is another retard - except this one is useful. The V1 generator required A LOT of manual work; V2 reads the pspsdk psplibdoc.xml file containing all the modules, NIDs, and names of everything in the PSP system, finds where the functions are defined in the sdk headers, and extracts the declaration. It then uses this info to build a function declaration that works with the BIOS (converting types, etc) and generates entire module files that can be compiled without any touching. Best of all it can version itself by storing versions in the files it generates and comparing them on the next run. It will only ever write out a file if the version is different (new methods added, etc) and in that case it writes it side-by-side with the original. Unfortunately some manual work is required there, but I'll live ;)

The FastHLE BIOS has nothing implemented yet. I just ran GenerateStubsV2 and it spat out those module classes, and I added just enough structure to get things compiling. Since the old BIOS is broken and the new one isn't done, that means that the emulator doesn't work! Ahhhhh! We're all gonna die!!!