More BIOS routines implemented - this time some of ThreadManForUser and other misc ones that were needed for the samples.
Fixed the bug that was causing crashes on shutdown - turns out that thread abort exceptions are ignored on threads executing in native code, and in the case of the samples that stay inside of native code this mean that shutdown would proceed without the threads (and thus the code execution) stopping. I have a check in there now on every syscall that sees if the app is shutting down and dies if it is. Nasty, but it has to be done.

Sample 1-5 (up to NeHe lesson 3) run. The rest don't run because I don't have any file or control methods implemented.

The bug in timing still exists - sometimes the app runs fine, other times it runs out of control. Very annoying...