Refactored some stuff in ICpu (got rid of the memory pointer stuff, as that belongs in IMemory, and added a SetupGame function).
On BREAK, the current address/instruction and the registers will be printed out.

SimpleElfDebugData added - this simply pulls the symbols from the ELF and gets the function names. It isn't really designed to be used by a debugger (that is what ElfDebugData will be for - loading code, etc).

The ultra CPU uses SimpleElfDebugData with a new TRACESYMBOLS option to output function names in to the trace log. This really helps with figuring out program flow.

PSPTris is dying horribly in libpng still. I'm having trouble diagnosing the issue - there are two, and they may be related. The first is that the CRC of the IHDR block fails - this means that either the data is bad, or crc32 (from zlib) is returning bad results. The second is that fread seems to be doing something retarded on certain calls:
- ftell() = position 16
- fread( 13 more bytes ) = read 13 bytes
- ftell() = position 13 !
What's weird is that it seems to work most of the time (even on previous 13 byte reads, so I don't think that's a magic number).
I'll keep adding diagnostic information until I can figure this out. I'm hoping I can avoid setting up a newlib build environment (as that seems messy), but it may be needed to get more debug info.