Native shim generation now implemented. I think the MT19937 RNG is the only thing in the bios using it.
Textures temporarily disabled - with this, psptris runs to the menu. sceCtrlReadLatch is not yet implemented, so I can't do anything, but it displays stuff. Since I haven't tried it on a real PSP I don't know what it's supposed to be doing, but from what it looks like it's the 'psp tris' logo and a bunch of random particle effects.
Textures aren't working probably because they are not 32bit RGBA (the only thing supported). Things run at about 137 FPS in debug mode and would probably be about twice as fast.

Some interesting stats discussion:
InstructionsExecuted: 1241354202
ExecutionLoops: 978 <-- way too high!
AverageCodeBlockLength: 36 <-- maybe I can bring the max length down from 200 if the avg is this low
JumpBlockLookupCount: 1194 <-- see note below
JumpBlockInlineHits: 115206001 <-- nice!
ManagedMemoryReadCount: 126234 <-- wtf?
AverageCodeSizeRatio: 13.2359290314544 <-- decent, considering that all the stats stuff is in there
AverageGenerationTime: 0.0153717598061174 <-- WAYYY too high

Note: that's a lot of lookups, and right below the hit count shows how many are hit (compared to misses, which is like 1000). It would be cool to do some sort of branch prediction by emitting a few nops, then after the first branch inlining a conditional jump to the target block, skipping the lookup.