Implemented FPL/VPL in FastHLE.
!! breaking changes !! emulator is currently broken!

Refactored a lot of the core library. Moving the loading code out, so added ILoader in the BIOS to do it for the host. Report generation and such is still in the library, but actual loading isn't. ElfFile is used by the debugging stuff still, so it's been moved in there.

Removed a bunch of cruft, such as the ObjdumpDebugData, IClock and IDmaController CPU classes, and other things.

Added XML documentation comments to everything in the core library. I was bored ^_^

Added a new BIOS - NativeHLE. FastHLE will have lived a short life. I really need to get as much of the BIOS native as possible, and by having all of the core BIOS types managed this couldn't happen. The new BIOS will have all native types (Kernel, Thread, etc) as well as a finely-tuned interrupt/timer/context switch system.
None of this is actually implemented and it doesn't compile, but ah well...

Note that FastHLE has not been updated after the refactor to work. I don't want to waste the time.