Getting things going.
Now most of the CPU threading code in place, as well as the proper support for things like the loader thread. I uncreatively called the mechanism a 'safety callback'. New threads get a dummy $ra value, and when they try to jr to them the CPU detects it. If the BIOS set a callback it will be executed, otherwise we assert (and should die). This allows the loader, for example, to create a thread, run the user module_start code, then kill itself when it tries to jr $ra. It seems to work great, besides the fact that the CPU cannot handle releasing the currently active thread.

Now I need to implement BIOS routines for everything. There is also the issue of instruction counts being totally wrong (I'm using some random register result instead of any sensical value).