Minor tweaks.
Fixed up TimerQueue so it works now. Not the most efficient thing in the world... The issue is that in the callback it needs to go from some unique timer ID to a timer as quickly as possible. Right now I am using a linked list and searching it for the timer. Assuming the timer count is low, it should be ok. There is a search counter in there to see how many iterations of the list are required. If this number gets big over the course of a session, it should be optimized.

PSPTris almost runs again - the 'PSPTris' logo bounces in!. Still a lot of module stuff to do (input, sceRtcs, MT1993).