PSPTris intro now runs for real.
New attributes added for bios functions - SuggestNativeAttribute and DontTraceAttribute. Suggest native is not used yet, but is there so that it will be easier to generate implementation reports. The idea is that functions that should be written in native code in the CPU should be marked with the attribute, and when those functions aren't implemented a warning can be generated. The DontTraceAttribute is used by the syscall debug printer to determine what calls should be ignored. This is better and easier to work with than the big if().

CPU perf is hurting (and has been forever) because of constant early returns after executing 10-100 instructions. The CPU should ONLY return when it makes a non-StatelessAttribute syscall or the stop flag is set.