Added the ability to have the logger print via Debug.WriteLine too, because in the case of the app dying the log is useless. Really need to have it trace to a file.
Played with textures - can't figure out what is wrong - using gDEBugger, everything is right - the textures show up on the first frame bound and everything. I _hate_ OpenGL.

Tried to implement resizing, but unfortunately to do it properly I need to set the projection matrix... Which is tricky because right now I'm just passing whatever the user code gives me straight to GL; I'd need to somehow decompose the projection matrix and change the viewport size. I'd also need to fix all of the sprite and texture transfer code. Lame.
The other option is to render to texture and then draw that surface larger. This would be the fastest way (still only drawing 480x272), but would suffer quality-wise because of the scaling. I'd much rather just draw everything at the desired native resolution.