Async IO implemented (except for close - so file handles are leaked now).
Implemented a faked msgdialog system - I don't know how I want to do it yet (interface that the frontend implements, maybe?), so I just dump it to the log and continue with what I think is the default answer.

With this, ToE runs almost to the menu (displays the scene animating in, then crashes when start is hit). The logo textures don't show up, either, possibly for the same reasons some textures don't show in PB.