Added a DONTTRACE define to IoFileMgrForUser.Files.cs - since it's nice to get rid of all the sceIoRead's and such but sometimes nice to have them for debugging.
Minor tweaks while trying to get things to run.
ToE is dying with some bad memory reads (at 0x0 and 0x4, so probably some base that is bad), but if those are ignored it gets further and tries to load some audio files or something - with a buffer address of 0x0...
Tried running another game, but it goes into a loop with savedata reads - need to fake that like I did the MsgDialog I think.