LL/LRU fixed (had some pretty glaring bugs). No longer leaking textures from LRU. LRU also now supports a entry free handler to properly cleanup entries on ejection.
Played with a bunch of texture checksum stuff - settled on an algo that checks every middle pixel of every 4 lines. This is probably retarded. I hope ector knows a better way. I had to do this because Tales of Eternia reuses texture addresses and does partial updates (so writing a cookie to the first word isn't enough). It seems fast enough - not lightning fast, but workable.

Now properly (kinda) emulate ambient material colors. This allows for fading and such that didn't work before.

Played with texture cache size. Of course larger is better, but it's amazing just how big the difference is. Right now it's defined as 1500, which seems to work well for Tales. In the future it would be nice to have this size change depending on how much memory is available. It would be even nicer to generalize the LRU to be based on per-entry amounts instead of count, so that a maximum amount could be defined and then each entry adds to it (in this case, estimated bytes occupied).

Now do CLUT checksums, and store the CLUT pointer/checksum used by a cached texture. This allows for textures to be regenerated when a CLUT changes - a common animation technique.