Fixes to texture loading and such. I think.
Fixed regression with textures in TrigWars/PSPTris - they were not showing up because the ambient material was getting set to black/0, but the pipeline was still drawing with it. I implemented LTE (Light Enable), but for some reason PB sets LTE=0, even though it needs the ambient material used. Need to figure out what's up.
So the current status is that TrigWars/PSPTris work again, and the fade on PB is gone for now.