Got rid of if()'s in the shader - until I find a nice tool that dumps assembly I don't know if the compiler was generating the most efficient code - the if's were cleaner, but this way it's explicit
Now cache uniform parameters in DefaultProgram - in the steady state there are very few glUniform calls now

StateManagement stuff reworked to be a bit better - now used for setting render state, arrays, etc.

Still disabling depth test for transformed primitives - probably just an issue with orthoMatrix - it has a depth range of -1,1 instead of the real range. The old plugin had the same issue, but now that I know what it is I might as well fix it... later.

I was hoping to be able to do the color widening in the shader, but it looks like it isn't practical - the GPU sucks at integer arithmetic.